+++ commit/reader_test.go
@@ -1,0 +1,4 @@
+// Copyright (c) Roman Atachiants and contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for details.
+
+package commit
